# [swift-4.0-branch][stdlib] Implement FixedWidthInteger.init<T : BinaryFloatingPoint>