* Explanation: This patch implements a generic conversion from any `BinaryFloatingPoint` type to any `FixedWidthInteger`
* Scope of Issue: Additive
* Risk: Minimal
* Reviewed By: Max Moiseev
* Testing: Automated test suite + test cases covering the new functionality
* Directions for QA: N/A
* Radar: <rdar://problem/34036943>