@moiseev 2017-08-30
@swift-ci Please test