@moiseev 2017-09-15
With the last commit reverting the constraints on protocols, here are the compilation time stats:


old | new | delta_pct | name
-- | -- | -- | --
66574 | 66728 | 0.23 | Sema.NumConformancesDeserialized
9400 | 9417 | 0.18 | Serialization.NumNormalProtocolConformancesLoaded
34709 | 34742 | 0.1 | Constraint solver overall.OverallNumSimplifiedConstraints
1388 | 1389 | 0.07 | Serialization.NumNormalProtocolConformancesCompleted
117617 | 117680 | 0.05 | Generic signature builder.NumArchetypeAnchorCacheHits
24916 | 24925 | 0.04 | Generic signature builder.NumConcreteTypeConstraints
808416 | 808746 | 0.04 | Generic signature builder.NumConformanceConstraints
571559 | 571805 | 0.04 | Generic signature builder.NumConformances
11187 | 11190 | 0.03 | Generic signature builder.NumArchetypeAnchorCacheMisses
407689 | 407785 | 0.02 | Generic signature builder.NumPotentialArchetypes
86319 | 86330 | 0.01 | Constraint solver overall.TotalNumTypeVariables
166539264 | 166559744 | 0.01 | Driver.ChildrenMaxRSS
530478 | 530520 | 0.01 | Generic signature builder.NumSameTypeConstraints
114870601 | 114861033 | -0.01 | AST.NumASTBytesAllocated
31712 | 31708 | -0.01 | Sema.NumLazyGenericEnvironments
31712 | 31708 | -0.01 | Serialization.NumLazyGenericEnvironments
7793 | 7790 | -0.04 | Serialization.NumMemberListsLoaded
214106 | 213947 | -0.07 | Sema.NumDeclsDeserialized
214106 | 213947 | -0.07 | Serialization.NumDeclsLoaded
22122 | 22020 | -0.46 | Name lookup.NumLazyIterableDeclContexts
22122 | 22020 | -0.46 | Sema.NumLazyIterableDeclContexts
12605 | 12506 | -0.79 | Name lookup.NumUnloadedLazyIterableDeclContexts
12605 | 12506 | -0.79 | Sema.NumUnloadedLazyIterableDeclContexts

And the benchmark results, that, honestly, don't make sense.

<details open>
  <summary>Regression (20)</summary>
  
TEST                                                    | OLD    | NEW    | DELTA  | SPEEDUP   
---                                                     | ---    | ---    | ---    | ---       
ObjectiveCBridgeStubToNSDate                            | 12879  | 17795  | +38.2% | **0.72x** 
ObjectiveCBridgeStubFromArrayOfNSString                 | 30068  | 38985  | +29.7% | **0.77x** 
SevenBoom                                               | 1396   | 1727   | +23.7% | **0.81x** 
ObjectiveCBridgeStubFromNSString                        | 980    | 1198   | +22.2% | **0.82x** 
NSError                                                 | 263    | 320    | +21.7% | **0.82x** 
ObjectiveCBridgeFromNSArrayAnyObjectToStringForced      | 39315  | 47201  | +20.1% | **0.83x** 
ObjectiveCBridgeFromNSArrayAnyObjectToString            | 45897  | 54091  | +17.9% | **0.85x** 
ArrayValueProp4                                         | 7      | 8      | +14.3% | **0.88x** 
ObjectiveCBridgeFromNSSetAnyObjectForced                | 4260   | 4836   | +13.5% | **0.88x** 
DropLastCountableRangeLazy                              | 8      | 9      | +12.5% | **0.89x** 
ObjectiveCBridgeFromNSDictionaryAnyObjectForced         | 6492   | 7234   | +11.4% | **0.90x** 
ObjectiveCBridgeFromNSDictionaryAnyObjectToString       | 129539 | 144166 | +11.3% | **0.90x** 
DropLastArray                                           | 10     | 11     | +10.0% | **0.91x** 
ObjectiveCBridgeFromNSDictionaryAnyObjectToStringForced | 117152 | 127101 | +8.5%  | **0.92x** 
ArrayInClass                                            | 79     | 85     | +7.6%  | **0.93x** 
PrefixWhileAnyCollection                                | 120    | 129    | +7.5%  | **0.93x** 
ObjectiveCBridgeStubNSDataAppend                        | 2330   | 2491   | +6.9%  | **0.94x (?)** 
ArrayAppend                                             | 690    | 737    | +6.8%  | **0.94x** 
StringWithCString                                       | 59938  | 63687  | +6.3%  | **0.94x** 
ObjectiveCBridgeFromNSDictionaryAnyObject               | 127043 | 134493 | +5.9%  | **0.94x** 

</details>

<details >
  <summary>Improvement (16)</summary>
  
TEST                                                    | OLD    | NEW    | DELTA  | SPEEDUP   
---                                                     | ---    | ---    | ---    | ---       
ReversedDictionary                                      | 85     | 68     | -20.0% | **1.25x** 
StringBuilder                                           | 1098   | 902    | -17.9% | **1.22x** 
ObjectiveCBridgeFromNSSetAnyObjectToString              | 86922  | 73813  | -15.1% | **1.18x** 
StringUTF16Builder                                      | 1951   | 1684   | -13.7% | **1.16x** 
ObjectiveCBridgeFromNSSetAnyObjectToStringForced        | 79867  | 68952  | -13.7% | **1.16x** 
ArrayValueProp                                          | 8      | 7      | -12.5% | **1.14x** 
ArrayValueProp2                                         | 8      | 7      | -12.5% | **1.14x** 
DictionaryRemove                                        | 2681   | 2391   | -10.8% | **1.12x (?)** 
NSStringConversion                                      | 979    | 879    | -10.2% | **1.11x** 
SetExclusiveOr_OfObjects                                | 9368   | 8592   | -8.3%  | **1.09x** 
SetExclusiveOr                                          | 3171   | 2919   | -7.9%  | **1.09x** 
RangeAssignment                                         | 348    | 321    | -7.8%  | **1.08x** 
ClassArrayGetter                                        | 13     | 12     | -7.7%  | **1.08x** 
ObjectiveCBridgeFromNSSetAnyObject                      | 71450  | 66422  | -7.0%  | **1.08x** 
FilterEvenUsingReduceInto                               | 144    | 135    | -6.2%  | **1.07x** 
HashTest                                                | 1626   | 1529   | -6.0%  | **1.06x** 

</details>