@moiseev 2017-09-15
@graydon I believe with the stats like this it should not be a problem to merge this. I just need to look closer at the benchmarks...
Also /cc @airspeedswift