@moiseev 2017-09-16
Here is a slightly better picture.

Re-running just the most regressed/improved benchmarks shows no difference.

<details open>
  <summary>Regression (3)</summary>
  
TEST                                               | OLD   | NEW   | DELTA  | SPEEDUP 
---                                                | ---   | ---   | ---    | ---     
ObjectiveCBridgeStubFromNSString                   | 981   | 1279  | +30.4% | **0.77x** 
NSError                                            | 270   | 310   | +14.8% | **0.87x** 
ObjectiveCBridgeFromNSSetAnyObjectForced           | 4480  | 4806  | +7.3%  | **0.93x** 

</details>

<details >
  <summary>Improvement (20)</summary>
  
TEST                                               | OLD   | NEW   | DELTA  | SPEEDUP 
---                                                | ---   | ---   | ---    | ---     
ObjectiveCBridgeStubFromNSStringRef                | 270   | 162   | -40.0% | **1.67x** 
ObjectiveCBridgeFromNSArrayAnyObjectToString       | 60170 | 48202 | -19.9% | **1.25x** 
ObjectiveCBridgeFromNSArrayAnyObjectToStringForced | 52578 | 42992 | -18.2% | **1.22x** 
ObjectiveCBridgeFromNSStringForced                 | 3026  | 2500  | -17.4% | **1.21x** 
SuperChars                                         | 87757 | 73843 | -15.9% | **1.19x** 
DictionaryRemove                                   | 2651  | 2311  | -12.8% | **1.15x** 
NSStringConversion                                 | 1007  | 878   | -12.8% | **1.15x** 
AnyHashableWithAClass                              | 67511 | 59920 | -11.2% | **1.13x** 
ObjectiveCBridgeStubFromArrayOfNSString            | 39517 | 35430 | -10.3% | **1.12x** 
ObjectiveCBridgeFromNSDictionaryAnyObjectForced    | 6628  | 5956  | -10.1% | **1.11x** 
DropLastArray                                      | 11    | 10    | -9.1%  | **1.10x** 
DropLastArrayLazy                                  | 11    | 10    | -9.1%  | **1.10x** 
StringBuilder                                      | 988   | 903   | -8.6%  | **1.09x** 
ErrorHandling                                      | 2753  | 2566  | -6.8%  | **1.07x** 
HashTest                                           | 1632  | 1537  | -5.8%  | **1.06x** 
ObjectiveCBridgeToNSSet                            | 45535 | 42887 | -5.8%  | **1.06x** 
DictionaryBridge                                   | 3711  | 3498  | -5.7%  | **1.06x** 
ObjectiveCBridgeFromNSSetAnyObject                 | 69004 | 65423 | -5.2%  | **1.05x** 
StringUTF16Builder                                 | 1816  | 1725  | -5.0%  | **1.05x** 
ObjectiveCBridgeToNSDictionary                     | 56889 | 54079 | -4.9%  | **1.05x** 

</details>